Put documents routes in its own file
